Transfer function to lib folder
